Create structuring-for-speaker-support-page-widget
Only looks at first "Description" in session file.

If the value includes the text "Zoom" then output is "Zoom". If the value includes the text "Embedded" then the output is "Embedded." If the text includes "Pre-recorded" then the output is "Embedded Simulive". If the text includes "Livestream" then the output is "Livestream via Zoom"